change rule repeat bug

+++ console/services/app_config/port_service.py
@@ -378,22 +378,22 @@ class AppPortService(object):
                 region_id = region.region_id
                 domain_repo.create_service_domains(service_id, service_name, domain_name, create_time, container_port, protocol,
                                                    http_rule_id, tenant_id, service_alias, region_id)
-                # 给数据中心发请求添加默认域名
-                data = dict()
-                data["domain"] = domain_name
-                data["service_id"] = service.service_id
-                data["tenant_id"] = tenant.tenant_id
-                data["tenant_name"] = tenant.tenant_name
-                data["protocol"] = protocol
-                data["container_port"] = int(container_port)
-                data["http_rule_id"] = http_rule_id
-                try:
-                    region_api.bind_http_domain(service.service_region, tenant.tenant_name, data)
-                except Exception as e:
-                    logger.exception(e)
-                    domain_repo.delete_http_domains(http_rule_id)
-                    return 412, "数据中心添加策略失败"
-
+                if service.create_status == "complete":
+                    # 给数据中心发请求添加默认域名
+                    data = dict()
+                    data["domain"] = domain_name
+                    data["service_id"] = service.service_id
+                    data["tenant_id"] = tenant.tenant_id
+                    data["tenant_name"] = tenant.tenant_name
+                    data["protocol"] = protocol
+                    data["container_port"] = int(container_port)
+                    data["http_rule_id"] = http_rule_id
+                    try:
+                        region_api.bind_http_domain(service.service_region, tenant.tenant_name, data)
+                    except Exception as e:
+                        logger.exception(e)
+                        domain_repo.delete_http_domains(http_rule_id)
+                        return 412, "数据中心添加策略失败"
         else:
             service_tcp_domains = tcp_domain.get_service_tcp_domains_by_service_id_and_port(
                 service.service_id, deal_port.container_port)
@@ -419,20 +419,21 @@ class AppPortService(object):
                 region_id = region.region_id
                 tcp_domain.create_service_tcp_domains(service_id, service_name, end_point, create_time, container_port,
                                                       protocol, service_alias, tcp_rule_id, tenant_id, region_id)
-                port = end_point.split(":")[1]
-                data = dict()
-                data["service_id"] = service.service_id
-                data["container_port"] = int(container_port)
-                data["ip"] = "0.0.0.0"
-                data["port"] = int(port)
-                data["tcp_rule_id"] = tcp_rule_id
-                try:
-                    # 给数据中心传送数据添加策略
-                    region_api.bindTcpDomain(service.service_region, tenant.tenant_name, data)
-                except Exception as e:
-                    logger.exception(e)
-                    tcp_domain.delete_tcp_domain(tcp_rule_id)
-                    return 412, "数据中心添加策略失败"
+                if service.create_status == "complete":
+                    port = end_point.split(":")[1]
+                    data = dict()
+                    data["service_id"] = service.service_id
+                    data["container_port"] = int(container_port)
+                    data["ip"] = "0.0.0.0"
+                    data["port"] = int(port)
+                    data["tcp_rule_id"] = tcp_rule_id
+                    try:
+                        # 给数据中心传送数据添加策略
+                        region_api.bindTcpDomain(service.service_region, tenant.tenant_name, data)
+                    except Exception as e:
+                        logger.exception(e)
+                        tcp_domain.delete_tcp_domain(tcp_rule_id)
+                        return 412, "数据中心添加策略失败"
 
         deal_port.is_outer_service = True
         if service.create_status == "complete":
